% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionConteosAcumulados.R
\name{funcionConteosAcumulados}
\alias{funcionConteosAcumulados}
\title{funcionConteosAcumulados}
\usage{
funcionConteosAcumulados(datos, bloques, nombre)
}
\description{
funcionConteosAcumulados
}
