% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionNivNetos.R
\name{funcionNivNetos}
\alias{funcionNivNetos}
\title{Función para frecuencias y pruebas de proporciones}
\usage{
funcionNivNetos(nNdatos, nNVars)
}
\description{
Función para frecuencias y pruebas de proporciones
}
